Compile ES9218 driver instead of ES9018
it was set to compile the for v10 dac(ES9018), the ES9218 is the v20 dac, though the es9018.c file does have some code
for the v20 dac(ES9218) so does this affect much?  ¯\_(ツ)_/¯ perhaps more features/functionality for the v20 dac as the
es9218.c should be more tailored to its dac. off the bat it sounds a bit bass-ier but more testing is needed

  Changes to be committed:
	modified:   arch/arm64/configs/lineageos_us996_defconfig
	modified:   sound/soc/codecs/Kconfig
	modified:   sound/soc/codecs/Makefile
	modified:   sound/soc/codecs/wcd-mbhc-v2.c
	modified:   sound/soc/codecs/wcd9335.c
	modified:   sound/soc/msm/msm8996.c