Elsa: force kernel to set dac to Hi-fi(1) when headphones are plugged in
The amp isnt being turned on at __es9218_sabre_headphone_on() or wcd2bypass resulting in hifi dac staying in BYPASS mode
so add code to make sure the es9218_amp on before it executes the code that will put it into hi-fi mode 1.
Now the dac enters hifi(1) mode (<50ohm) but wo't power off/go into bypass after headphones are unplugged in function es9218_sabre_amp_stop (the g_headset_type is 0 and says its invalid)
so make the default case: hi-fi1->BYPASS this last part is very hacky so pick at your own risk

AFAIK es9218_headset_type_get sets the value for the headset to be the dac and amp can pick high-impedance/aux/normal modes
but for my device(or maybe in lineage/aosp in general) the value doesnt change from 0 (maybe just (>50ohm headphones)). resulting in the dac staying in bypass mode as the bypass2hifi function
will not work and say 0 is invalid. So this is my workaround for now, if the rest of the code is working fine it could be as easy as fixing it so that it assigns a value
to the g_headset_type on its own, another solution is to add sys_fs nodes to control this value but this is just in theory.

To monitor the dac in kernel you can use dmesg to see the code working with a terminal app with root access

dmesg -w| grep -iC 0 es9* (the -w makes it continuious if you dont want that delete it as youll probably have to force-close the app)

you can also monitor in tinymix:

tinymix | grep Es9*

Es9018 state: Hifi (means its on)
Es9018 state: Bypass (? a primed mode ready to be turned on?)
Es9018 state: Close (off/sleep no signal from the dac when theres no headphones plugged in.)

To reiterate its probably the g_headset_type being 0 causing this and this may only work around that rather than fixing the real problem. but idk ¯\_(ツ)_/¯

 Changes to be committed:
	modified:   sound/soc/codecs/es9218.c